third_party/blink/web_tests/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/parallel-compute.https.any.worker_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] test parallel WebNN API compute operations
  assert_array_equals: expected property 1 to be 4 but got 6 (expected array [2, 4, 6, 8] got [2, 6, 8, 8])
Harness: the test ran to completion.
